% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bCI_tab.R
\name{bCI_tab}
\alias{bCI_tab}
\title{CI compute}
\usage{
bCI_tab(b, ind = length(b$t0), type = "perc", conf = 0.95)
}
\arguments{
\item{boo01}{bootMer object}
}
\value{

}
\description{
Compute 95\% confidence interval of a mixed model
}
\examples{
boo01 <- lme4::bootMer(m2, mySumm, nsim = 2, re.form = NA)
bCI.tab(boo01)

}
