% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCIcordifG29G7.R
\name{computeCIcordifG29G7}
\alias{computeCIcordifG29G7}
\title{CI difference of correlations}
\usage{
computeCIcordifG29G7(pair = "Cherry_Cranberry")
}
\arguments{
\item{pair}{pair of environments}
}
\value{

}
\description{
Compute 95\% confidence interval of the difference between two correlations
}
\examples{
computeCIcordifG29G7(pair="Cherry_Cranberry")
}
