% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCIcordifG7G29.R
\name{computeCIcordifG7G29}
\alias{computeCIcordifG7G29}
\title{CI difference of correlations}
\usage{
computeCIcordifG7G29(pair = "Cherry_Cranberry")
}
\arguments{
\item{pair}{pair of environments}
}
\value{

}
\description{
Compute 95\% confidence interval of the difference between two correlation coefficient based on Zou 2007 (Zou, G. Y. (2007). Toward using confidence intervals to compare correlations. Psychological methods, 12(4), 399.)
}
\examples{
computeCIcordifG29G7(pair="Cherry_Cranberry")
}
