% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecorrelation.R
\name{computecorrelation}
\alias{computecorrelation}
\title{Correlation Function}
\usage{
computecorrelation(data)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Compute correlation between two variables
}
\examples{
ddply(data_logchangeNb_eggsEmergence_rate, .(Generation, Fruit_s), computecorrelation)
}
