% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computefitness.R
\name{computefitness}
\alias{computefitness}
\title{Calculation fitness during phenotyping steps}
\usage{
computefitness(
  fitness_dataset_intermediate = data_G7,
  fitness_dataset_final = data_G29
)
}
\arguments{
\item{fitness_dataset_intermediate}{Intermediate phenotyping dataset created using loadfitnessdata()}

\item{fitness_dataset_final}{Final phenotyping dataset created using loadfitnessdata()}
}
\value{
dataset of fitness difference between the phenotyping step and the initial phenotyping step
}
\description{
Fitness during phenotyping step: w = log(mean/20)
}
\examples{
data_sum <- computefitness(fitness_dataset_intermediate = data_G7, fitness_dataset_final = data_G29)
}
