% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computelogchange_forlongdata.R
\name{computelogchange_forlongdata}
\alias{computelogchange_forlongdata}
\title{Calculation difference in fitness: from longitudinal data to logchange}
\usage{
computelogchange_forlongdata(longitudinal_dataset = data_sum)
}
\arguments{
\item{longitudinal_dataset}{Longitudinal dataset created using loadlongitudinaldata()}
}
\value{
dataset of fitness difference between fitness during step and phase 1 mean per fruit
}
\description{
Log change in fitness in the three phase for the population i in selective fruit j

= log(mean(nb_adults_ij_allgenerationphase1/mean(nb_adlts_j_phase1))
= log(mean(nb_adults_ij_allgenerationphase3/mean(nb_adlts_j_phase1))
Example for CE2 during phase1
log(mean (CE2 in G2 G3 G4 G5) / mean (all lines Cherry in phase1))

Mean and sd are weighted by the number of tubes counted
}
\examples{
data_sum <- computelogchange_forlongdata(longitudinal_dataset = data_sum)
}
