% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computelogchange_POOL_hypo1.R
\name{computelogchange_POOL_hypo1}
\alias{computelogchange_POOL_hypo1}
\title{Calculation difference in fitness: from fitness data to logchange for hypo 1}
\usage{
computelogchange_POOL_hypo1(
  fitness_dataset_intermediate = data_G7,
  fruit1 = "Cherry",
  fruit2 = "Cranberry"
)
}
\arguments{
\item{fitness_dataset_intermediate}{Intermediate phenotyping dataset created using loadfitnessdata()}

\item{fruit1}{First fruit of the pairwise}

\item{fruit2}{Second fruit of the pairwise}
}
\value{
Dataset with logchange for two fruits during a specific phenotyping step
}
\description{
Log change in fitness in generation x for the fruit sel i in the test fruit j = mean(W)Gxij / mean(W)G1j

Log change:

logchange=Ln(x1/x2) with x1 and the x2 the mean of Nt+1 and Nt respectively

logchange=ln((meanNt+1)/mean(Nt))

sd log change with Hedges et al. 1999

se(logchange)=sqrt((s1²/n1x1²)+(s2²/n2x2²)) with s1 and s2 the standard deviation of Nt+1 and Nt respesctively,

and n1 and n2 the sample size of  Nt+1 and Nt respesctively

In our case: x2=N0 and s2=sd N0

se(logchange)=sqrt((s1²/n1x1²))=sqrt((sdNt+1²/NNt+1 x meanNt+1²)+(sdN0²/N0 x meanN0²))
}
\examples{
data_sum <- computelogchange_POOL_hypo1(fitness_dataset_intermediate = data_G7, fruit1 = "Cherry",fruit2 = "Cranberry")
}
