% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computemeanvar_fitnessdata.R
\name{computemeanvar_fitnessdata}
\alias{computemeanvar_fitnessdata}
\title{Compute the mean and variance in fitness using bootstrapping}
\usage{
computemeanvar_fitnessdata(seed = 1, ntubes = 30, N = 25, theta = 3)
}
\arguments{
\item{theta}{}
}
\value{

}
\description{
Compute the mean and variance in fitness using bootstrapping
}
