% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_r.R
\name{eq_r}
\alias{eq_r}
\title{Create formatted equation for ggplot}
\usage{
eq_r(gen = 7, pair = "Cherry_Cranberry")
}
\arguments{
\item{gen}{Phenotyping step of interest}

\item{pair}{vector with first fruit of the pairwise and second fruit of the pairwise}
}
\value{
Formatted equation
}
\description{
Create formatted equation for ggplot
}
\examples{
eq_r(gen = 7, pair = "Cherry_Cranberry")
}
