% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_r_ci.R
\name{eq_r_ci}
\alias{eq_r_ci}
\title{Create formatted equation for ggplot}
\usage{
eq_r_ci(gen = 7, pair = "Cherry_Cranberry")
}
\arguments{
\item{gen}{Phenotyping step of interest}

\item{pair}{vector with first fruit of the pairwise and second fruit of the pairwise}
}
\value{
Formatted equation
}
\description{
Create formatted equation for ggplot
}
\examples{
eq_r_ci(gen = 7, pair = "Cherry_Cranberry")
}
