% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filldata_maregression.R,
%   R/filldata_predict_maregression.R
\name{filldata_maregression}
\alias{filldata_maregression}
\title{Fill out an empty dataset with prediction from MA-regression}
\usage{
filldata_maregression(
  model = mod0_G7_CheCran,
  empty_dataset = Summary_estimates_MAregression,
  generation = 7,
  pairwise = "Cherry_Cranberry"
)

filldata_maregression(
  model = mod0_G7_CheCran,
  empty_dataset = Summary_estimates_MAregression,
  generation = 7,
  pairwise = "Cherry_Cranberry"
)
}
\arguments{
\item{model}{output model using lmodel2() ex: mod0_G7_CheCran}

\item{empty_dataset}{empty dataframe ex. Summary_estimates_MAregression}

\item{generation}{generation in the model: 7 or 29}

\item{pairwise}{pairwise of fruit: Cherry_Cranberry, Cranberry_Strawberry, Strawberry_Cherry}
}
\value{



}
\description{
Fill out empty dataset with prediction from Ma-regression

Fill out empty dataset with prediction from Ma-regression
}
\examples{
Estimates_pairwise <- filldata_maregression(model = mod0_G7_CheCran, empty_dataset = Estimates_pairwise, generation = 7, pairwise = "Cherry_Cranberry")

Summary_estimates_MAregression <- Summary_estimates_MAregression(model = mod0_G7_CheCran, empty_dataset = Summary_estimates_MAregression, generation = 7, pairwise = "Cherry_Cranberry")

}
