% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_limits.R
\name{filldata_predict_maregression}
\alias{filldata_predict_maregression}
\title{Calculate polygon limits from lmodel2}
\usage{
filldata_predict_maregression(model = mod0_G7_CheCran)
}
\arguments{
\item{model}{output model using lmodel2()}
}
\value{

}
\description{
Calculate polygon limits for Ggplot2 (grey area)
}
\examples{
polygondata <- filldata_predict_maregression(model = mod0_G7_CheCran)
}
