% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillout_estimate_correlation.R
\name{fillout_estimate_correlation}
\alias{fillout_estimate_correlation}
\title{Fill out dataset with correlation estimates}
\usage{
fillout_estimate_correlation(
  emptydata = Estimates_pairwise,
  gen = 7,
  pair = "Cherry_Cranberry"
)
}
\arguments{
\item{emptydata}{Empty dataset}

\item{gen}{Phenotyping step of interest}

\item{pair}{vector with first fruit of the pairwise and second fruit of the pairwise}
}
\value{
Filled dataset
}
\description{
Fill out dataset with estimates from correlation test and lmodel2 output
}
\examples{
Estimates_pairwise <- fillout_estimate_correlation (Estimates_pairwise, 7, "Cherry_Cranberry")
}
