% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_POOL_hypo1.R
\name{formatting_POOL_hypo1}
\alias{formatting_POOL_hypo1}
\title{HYPOTHESIS 1}
\usage{
formatting_POOL_hypo1(fitness_dataset_intermediate = data_G7)
}
\arguments{
\item{fitness_dataset_intermediate}{Intermediate phenotyping dataset created using loadfitnessdata()t}
}
\value{

}
\description{

}
\examples{
data_sum <- formatting_POOL_hypo1(logchange_dataset = data_logchange)
}
