% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattinglogchange.R
\name{formattinglogchange}
\alias{formattinglogchange}
\title{Formatting a logchange dataset to pairwise dataset for testing MA-regression}
\usage{
formattinglogchange(
  logchange_dataset = data_logchange,
  generation = "7",
  fruit1 = "Cherry",
  fruit2 = "Cranberry"
)
}
\arguments{
\item{generation}{Phenotyping step of interest}

\item{fruit1}{First fruit of the pairwise}

\item{fruit2}{Second fruit of the pairwise}

\item{logchange_data}{Intermediate phenotyping dataset created using loadfitnessdata()}
}
\value{
Dataset with logchange for two fruits during a specific phenotyping step
}
\description{
Formatting a logchange dataset
From a logchange dataset created using computelogchange() to pairwise dataset for testing MA-regression
}
\examples{
data_sum <- formattinglogchange(logchange_dataset = data_logchange, generation = "7", fruit1 = "Cherry",fruit2 = "Cranberry")
}
