% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfitnessdata.R
\name{loadfitnessdata}
\alias{loadfitnessdata}
\title{Title}
\usage{
loadfitnessdata(
  dataset = "Selection_Phenotypage_G0_G7_G8.csv",
  generation = "G1"
)
}
\value{

}
\description{
Title
}
\examples{

}
