% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_minmax.R
\name{vec_minmax}
\alias{vec_minmax}
\title{Create formatted vector with limits for ggplot}
\usage{
vec_minmax(data_set = data_logchange, gen = "7", fruit = "Cherry")
}
\arguments{
\item{data_set}{Dataset used for the plot}

\item{gen}{Phenotyping step of interest}

\item{fruit}{selective fruit}
}
\value{
Vector with minimum and maximum valuefor axis ggplot
}
\description{
Create formatted vector with minimum and maximum value for axis ggplot
}
\examples{
vec_minmax(data_logchange,"7","Cherry")
}
