% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fitnessdata.R
\name{boot_fitnessdata}
\alias{boot_fitnessdata}
\title{Bootstrap Function That Resample Fitness Date}
\usage{
boot_fitnessdata(seed, vecindpertube)
}
\arguments{
\item{seed}{Seed used of reproductibility of results}

\item{vecindpertube}{vector with fitness data}
}
\value{

}
\description{
Bootstrap Function That Resample Fitness Date
}
\examples{
bootfunc(seed=1, vecindpertube=sim_derived)
}
