% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lmodel2.R
\name{bootstrap_lmodel2}
\alias{bootstrap_lmodel2}
\title{Lmodel2 with bootstrap}
\usage{
bootstrap_lmodel2(
  seed = 1,
  data_pairwise = TEMP_dataG7_CheCran,
  sd_NA = mean_sd_G7
)
}
\arguments{
\item{data_pairwise}{dataset from pairwise of fruits, produces using formattinglogchange()}

\item{sd_NA}{average of the sd in all environments, to replace NA value}
}
\value{

}
\description{
Lmodel2 with bootstrap
}
\examples{
sim <- bootstrap_lmodel2(seed=1, data_pairwise = TEMP_dataG7_CheCran, sd_NA = mean_sd_G7)
}
