% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computelogchange_forlongdata.R
\name{computelogchange_forlongdata}
\alias{computelogchange_forlongdata}
\title{Calculation difference in fitness: from longitudinal data to logchange}
\usage{
computelogchange_forlongdata(longitudinal_dataset = data_sum)
}
\arguments{
\item{longitudinal_dataset}{Longitudinal dataset created using loadlongitudinaldata()}
}
\value{
dataset of fitness difference between fitness during step and G2 mean per fruit
}
\description{
Log change in fitness in generation x for the population i in the test fruit j = mean(W)Gxij / mean(W)G1j
}
\examples{
data_sum <- computelogchange_forlongdata(longitudinal_dataset = data_sum)
}
