% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computemeanvar_fitnessdata.R
\name{computemeanvar_fitnessdata}
\alias{computemeanvar_fitnessdata}
\title{Fitness Mean And Variance}
\usage{
computemeanvar_fitnessdata(seed = 1, ntubes = 30, N = 25, theta = 3)
}
\arguments{
\item{seed}{Seed used for simulation}

\item{ntubes}{Number of tubes whose individuals have been counted}

\item{N}{Mean of the negative binomial distribution (i.e. average number of individuals counted per tube)}

\item{theta}{Overdispersion parameter of the negative binomial distribution (i.e. of the count of individuals per tube)}
}
\value{

}
\description{
Compute the mean and the variance in fitness of a population using either the delta method or bootstrapping
}
\examples{
computemeanvar_fitnessdata(seed=1, ntubes=30, N=25, theta=3)
}
