% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computemeanvar_fitnessdifferencedata.R
\name{computemeanvar_fitnessdifferencedata}
\alias{computemeanvar_fitnessdifferencedata}
\title{Compute the mean and variance in fitness using bootstrapping}
\usage{
computemeanvar_fitnessdifferencedata(
  seed = 1,
  ntubes_ancestor = 30,
  N_ancestor = 25,
  theta_ancestor = 3,
  ntubes_derived = 30,
  N_derived = 30,
  theta_derived = 3
)
}
\arguments{
\item{theta}{}
}
\value{

}
\description{
Compute the mean and variance in fitness using bootstrapping
}
