% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_overdisp.R
\name{estim_overdisp}
\alias{estim_overdisp}
\title{Analyze A Dataset Using A Negative Binomial Distribution}
\usage{
estim_overdisp(fruit, data, generation)
}
\arguments{
\item{fruit}{Type of fruit considered}

\item{data}{Dataset considered}

\item{generation}{Generation considered}
}
\value{

}
\description{
Analyze a dataset using a negative binomial distribution and returns the fitted mean, variance and overdispersion
}
\examples{
estim_overdisp(data=data_G0, fruit="Cherry", generation="G0")

}
