% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattinglogchange.R
\name{formattinglogchange}
\alias{formattinglogchange}
\title{Formatting a logchange dataset to pairwise dataset for testing MA-regression}
\usage{
formattinglogchange(
  logchange_dataset = data_logchange,
  generation = "7",
  fruitcomb = c("Cherry", "Cranberry")
)
}
\arguments{
\item{logchange_dataset}{dataset with log fitness change between two phenotyping steps}

\item{generation}{Phenotyping step of interest}

\item{fruitcomb}{Pair of fruit media considered}
}
\value{
Dataset with logchange for two fruits during a specific phenotyping step
}
\description{
Formatting a logchange dataset
From a logchange dataset created using computelogchange() to pairwise dataset for testing MA-regression
}
\examples{
TEMP_dataG7_CheCran <- formattinglogchange(logchange_dataset = data_logchange, generation = "7", fruitcomb=c("Cherry", "Cranberry"))
}
