% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfitnessdata.R
\name{loadfitnessdata}
\alias{loadfitnessdata}
\title{Title}
\usage{
loadfitnessdata(
  dataset = "Selection_Phenotypage_G0_G7_G8.csv",
  generation = "G1"
)
}
\arguments{
\item{dataset}{Fitness dataset}

\item{generation}{Generation considered for the subset}
}
\value{

}
\description{
Title
}
\examples{
data_G29 <- loadfitnessdata(dataset = "PERFORMANCE_Comptage_adultes_G13G14G15G16G17G18G19G20G21G22G23G24G25G26G27G28G29.csv", generation = "29")
}
