% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfitnessdata_allfruits.R
\name{loadlongitudinaldata_allfruits}
\alias{loadlongitudinaldata_allfruits}
\title{Longitudinal fitness data Import Function for ALL FRUITS}
\usage{
loadlongitudinaldata_allfruits(
  dataset = "DATA_Adults_G1G29.csv",
  rm_generation1 = 1,
  rm_generation2 = 7,
  rm_generation3 = 29
)
}
\arguments{
\item{dataset}{Fitness dataset: csv file}

\item{rm_generation1}{First generation removed due to GF development or parental GF development}

\item{rm_generation2}{Second generation removed due to GF development or parental GF development}

\item{rm_generation3}{Third generation removed due to GF development or parental GF development}
}
\value{

}
\description{
Reads a fitness data file in table format and subset it to return the dataframe with all the generations
}
\examples{
data_sum <- loadlongitudinaldata_allfruits(dataset = "DATA_Adults_G1G29.csv", rm_generation1 = 1,rm_generation2 = 7,rm_generation3 = 29)
}
